refactor: collapse StateWriter interface into *GethWriter
With Erigon gone there is only one StateWriter implementation; the
interface added ~45 LOC of indirection without runtime polymorphism.

- `generator/writer.go` — **deleted entirely**.
- `generator/writer_geth.go` — absorb the pieces that survived:
  OutputFormat type + OutputGeth const (both vestigial, removed in the
  next commit), ParseOutputFormat (now trivial), WriterStats struct.
  Also delete the no-op GethWriter.WriteGenesisBlock that was only
  there for interface compliance — nothing calls it (the real work is
  done by genesis.WriteGenesisBlock at the package level, called from
  main.go and tests directly).
- `generator/generator.go`:
  - `Generator.writer` field changes from `StateWriter` to
    `*GethWriter`.
  - Drop the `StateWriter(gethWriter)` conversion in New.

All call sites (g.writer.Write*, Flush, SetStateRoot, Stats) keep
working unchanged because GethWriter exposes the same method set.

Tests: generator + genesis suites pass. Golden-hash tests unchanged.
