--- generator/writer.go
@@ -1,70 +1,0 @@
-package generator
-
-import (
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/params"
-)
-
-// OutputFormat specifies the database format to generate.
-type OutputFormat string
-
-const (
-	// OutputGeth generates a geth-compatible Pebble database (snapshot layer).
-	OutputGeth OutputFormat = "geth"
-)
-
-// ParseOutputFormat parses an output format string.
-// Retained as a vestigial shim while the OutputFormat type still exists
-// (removed in a follow-up commit). Unknown values fall through to geth.
-func ParseOutputFormat(s string) OutputFormat {
-	return OutputGeth
-}
-
-// StateWriter abstracts database writes for different client formats.
-type StateWriter interface {
-	// WriteAccount writes an account to the database.
-	// addrHash is the pre-computed keccak256(addr) to avoid redundant hashing.
-	WriteAccount(addr common.Address, addrHash common.Hash, acc *types.StateAccount, incarnation uint64) error
-
-	// WriteStorage writes a storage slot for an account.
-	// addrHash and slotHash are pre-computed keccak256 hashes so the writer
-	// doesn't hash redundantly; addr and slot are retained for external
-	// tooling that reads the raw keys.
-	WriteStorage(addr common.Address, addrHash common.Hash, slot common.Hash, slotHash common.Hash, value common.Hash) error
-
-	// WriteStorageRLP writes a storage slot with pre-encoded RLP value.
-	// Used when the caller already has the RLP encoding (avoids double-encode).
-	WriteStorageRLP(addrHash common.Hash, slotHash common.Hash, valueRLP []byte) error
-
-	// WriteRawStorage writes a storage slot using a pre-hashed trie key.
-	// The hashedSlot is used directly as the snapshot/trie key without
-	// keccak256 hashing. Used for phantom entries in deep-branch mode.
-	WriteRawStorage(addr common.Address, incarnation uint64, hashedSlot, value common.Hash) error
-
-	// WriteCode writes contract bytecode.
-	WriteCode(codeHash common.Hash, code []byte) error
-
-	// SetStateRoot sets the computed state root (for metadata/markers).
-	SetStateRoot(root common.Hash) error
-
-	// WriteGenesisBlock writes the genesis block and chain config.
-	// Only called when genesis integration is enabled.
-	WriteGenesisBlock(config *params.ChainConfig, stateRoot common.Hash) error
-
-	// Flush ensures all pending writes are committed.
-	Flush() error
-
-	// Close closes the writer and releases resources.
-	Close() error
-
-	// Stats returns write statistics.
-	Stats() WriterStats
-}
-
-// WriterStats holds statistics about database writes.
-type WriterStats struct {
-	AccountBytes uint64
-	StorageBytes uint64
-	CodeBytes    uint64
-}
